Add `enduire` binary
This isn't doing anything right now, but the plan is to have this be the
command you run locally when you want to snapshot something.

I had to temporarily give up my attempt at using async/await for now
(ran into babel issues that I couldn't resolve). But I'm hoping to get
to this later.