Use webpack to read test files
This is mostly a milestone commit. Tests are failing and things aren't
fully polished. It's the result after testing out multiple ways of
reading and executing snapshot files.

My attempt at using babel-register failed because I still needed webpack
for a few things. And it didn't play well with symlinks (which is a
prereq for working on enduire locally).

In the end, what worked best was to use webpack to construct a bundle.
The bundle is then executed and snapshot functions are stored on a
global snaps object.