Treat jsdom more like a browser
Instead of mixing execution of node and jsdom code, we're doing what the
jsdom readme suggests, which basically means treating jsdom as a
browser. Instead of requiring code directly in node, we `eval` it in the
jsdom environment. This doesn't save us from all the code injection
security issues but it's a step in the right path.

Running through eval was also a prerequisite for getting all happo
examples defined in the brigade web codebase to properly run.

Also included in this commit:
- a change to make the file include pattern configurable.
- a fix for examples returning a string instead of a react element.

https://github.com/tmpvar/jsdom#executing-scripts