Move to a worker-based model
Instead of synchronously waiting for a remote service to take our
snapshots and report back, we adopt a "ticketing" model where you
request a run, then wait for results. A worker will pick up your request
and report its result for you to fetch on the next tick.

This means we will now only talk to one API (instead of the two previous
ones).