Allow array structure for examples files
Instead of just depending on individual exports, we can allow example
files to export a list of components and its variants. This is useful if
you need to generate happo examples from a source in a different format.
In one project I work with, we define examples in a format that best
fits the style-guide. We then grab examples from this structure to form
happo examples.