Don't assume react-dom is in local node_modules folder
I was running into an issue when setting up happo.io for happo-view
where it couldn't find react-dom. It turns out we were hardcoding it to
be in a specific folder, which isn't always true (yarn will hoist
common dependencies to the root for instance). I'm actually not 100%
sure why we need an absolute path, but I tried without it (just
"require('react-dom')") and it didn't work.