Prevent duplicate React and ReactDOM in bundle
I'm running into an issue in one of my projects where I have a Lerna
structure. Happo is defined in the root and happo examples are allowed
at any place in the packages structure. When individual files require
React, they will use the closest one based on the package structure.
This means that multiple React copies might end up in the resulting
bundle, which leads to issues down the line. By setting react and
react-dom as externals (injected globally at the top), we should be able
to prevent the duplication.