Use url that uploadReport sends back
I've made some changes over at happo.io to improve meta tags for logged
out views of compare and report pages. To get this working properly, I
had to add a query param to the urls used to access these pages. To
ensure that this query param is used, the POST /api/reports endpoint now
returns the full url to a report.