Remove dependency on nodegit
This dependency turned out to be quite heavy in nature, requiring
binaries built for different OS'es, and with that all sorts of
installation issues.

We're not really depending on it other than generating a hash, which we
can do in a much cheaper way. The downside is that the generated token
isn't deterministic the same way the old sha was constructed. But I
think we can live with that.

An unrelated bugfix snuck in here as well:
Missing getRootElement in config would cause failures.