Add explicit dependency on request library
This is showing up as a warning when installing the happo.io lib:

warning "happo.io > request-promise-native@1.0.5" has unmet peer dependency "request@^2.34".
warning "happo.io > request-promise-native > request-promise-core@1.1.1" has unmet peer dependency "request@^2.34".

I've been saved by the fact that request is being included as part of a
different package (not sure which). But adding it should help prevent
weird issues as a result of it being missing. And users shouldn't have
to see this warning.