Always exit `happo compare` with a zero exit code
Before, we would use the exit code to signal whether two reports diff or
not. This led to some issues when the script is used in combination with
the Happo github app, which will post back statuses to the PR. The
happo endpoint is now more likely to break, which means exit code can no
longer be ignored. I'm updating things so that the output is empty when
there are no diffs. We can then use that as a signal in the CI script.