Use `npm run` in happo-ci
Before, we used to conditionally use yarn to run commands. Yarn run has
an annoying behavior where exit codes of inside scripts are not
propagated to the outside command. For a 113 exit code, `yarn run` will
exit with a 1. Npm on the other hand, correctly propagates exit codes.

We will still use yarn to install dependencies however, if we detect a
yarn.lock file.