Mock window dimensions in jsdom environment
Since we render components in a jsdom environment, then send the payload
(html + css) to happo.io, we've basically executed javascript in a 0x0
sized window. Any code relying on window dimensions would have had to
work with that limitation. This isn't ideal, and we can quite easily fix
it by overriding dimensions with the ones we have specified in the
.happo.js config.

The downside with this approach is that we have to render all components
multiple times. One for each target. The upside is that any code relying
on window dimensions (innerWidth, outerWidth, screen.width, etc) will
generate the right markup.