Stop eval-ing webpack bundle
I'm working on adding a plugin concept to the happo.io client. In order
to do that, I need a simpler main setup. The way we were sandboxing
jsdom and evaling the webpack bundle was adding unnecessary complexity,
so I'm moving back to a model where `document` is available globally in
node.

This essentially reverts 7d7573c. While I still think it's valuable to
sandbox js execution, the complexity cost is too high.