Properly clean out tmp file
I tried to make things work without having a locally stored tmp file,
but it turns out node's `require` function can't be made to lookup
dependencies local to a different place than the current file. It seems
Node 9 will have an options argument to pass to require, so in the
future we might be able to move the temp file back to os.tmpdir() and
then do

require(webpackBundle, { paths: [process.cwd()] });

Another option would be to not use a file at all. I might explore this
option further in the future.