Fix parallell builds
We were writing to the same files for the plain project and the react
project. Splitting them up will allow us to run tests in parallell. It
probably doesn't matter for the average user, but it might benefit some
to be able to run two different projects in the same cwd.