Add better support for portal rendered React components
It's not uncommon for a React component to render its content into a
different dom element than the current React root. Due to how we were
assuming that the root element was always the React target, we missed
some markup.

Here's a report from what we think is a portal rendered component:
https://happo.io/a/29/report/dev-e71f8f7a01084624e7c8

By rearranging the logic around how we find the root element, we can
add code that will fallback to a different root if the regular root is
free from content.