Make babel-core, babel-loader pure peer dependencies
It is assumed that most consumers will already have these in their
package.json folders, so making them peer isn't a big deal. The benefit
is that we allow better control for consumers, which can help prevent
duplicate versions of babel in their node_modules folders.

I thought about adding webpack here as well, but knowing very little
about webpack 4 I don't know if it's safe to allow ^3||^4 without
issues. And if we only add ^3, we make it really hard for users on
webpack 4.