Wait for keypress on dev change detected
Instead of automatically trying to generate screenshots every time the
user saves a file, we now wait for a keypress before we continue. This
is better because it allows you to work on the file for a while before
committing to generate the screenshots. Often times, there will be build
failures before you can continue anyway.