Improve logging around screenshooting
I tried a while to make the log editable, adding checks next to the
browser targets that were done. But that turned out to be hard when
other things can log things to stdout. Instead of adding code that takes
control over stdout printing completely, I'm simply making it log at the
end of the stream.