Use temporary pull-request token for PRs from forks
Providing secret environment variables in CI to forks is tricky. Most CI
environments, such as Travis, explicitly disable secrets when it detects
that a run is triggered on a fork of a repo.
https://github.com/travis-ci/travis-ci/issues/1946

To work around this, I've adopted a separate auth mechanism for happo.io
for outside pull requests. Instead of requiring an apiKey and an
apiSecret, we use temporary tokens, pulled from the Happo API if we
detect that we're in a fork (keys are missing). These tokens are much
more limited in scope compared to the regular API tokens. We're
basically allowing them to generate reports for the PR in question, and
only for a limited time.