Unmount React examples after render
Old happo used to do this. I ignored it for a while thinking it would
speed things up not unmounting things. It turns out large projects will
run into memory issues if they don't clean up after themselves.

For now, the cleanup phase of the plain js type is a no-op. It's likely
that this will change in the future (I plan on moving `type` into
plugins instead).