Make webpack a peer dependency to support webpack 4
Up until now, happo has bundled webpack with the library. This is
starting to be painful to consumers as they have to come up with clever
workarounds (#17). In the testing I've done, it looks like our usage of
webpack is compatible across webpack 3 and 4, so I'm simply moving the
webpack dependency to a peer one, allowing consumers to provide their
own copy. This is similar to what we did in 643bcd5 for babel
dependencies.