Further extract DomProvider, introduce puppeteer plugin
To better support projects that rely on DOM measurements, I've been
working on a dom provider using puppeteer (chrome). The result is
happo-plugin-puppeteer which can be used to replace JSDOM with chrome.

To test this, I made the plain-test use the new plugin, and rely on
measurements in one of the examples.