Update check for installed happo package
The previous check looked for a "happo" binary. I'm seeing this check
fail in Travis in the react-dates project:

https://travis-ci.org/airbnb/react-dates/jobs/435493766

It might be that nvm holds references to previous binaries which makes
our check fail. Instead of looking for the binary we can check if the
happo.io package is installed.

Yarn and npm does things slightly differently here. npm will output a
non-zero exit code if the package is missing. Yarn will not, so we have
to use grep trickery to make it behave.