Fix error handling when rendering examples
A recent update (7afc5b7d1) extracted DOM rendering into a separate
module. In the process, we changed error handling so that errors are
thrown from the webpack bundle instead of from within node. This meant
that our `instanceof` checking didn't work (because it was looking at a
serialized/deserialized object). We can mitigate this by adding a
property to the error that we throw.