Cut long jsdom errors
A customer is reporting happo logs larger than 100MB in CI. The main
reason for this is that jsdom fails to parse the css, then it console
logs that css. We can mitigate this by cutting long errors before we
send them to the node console.

To see the full log, people can set the `VERBOSE` env variable. This is
a hack and I've decided not to document this in the README (it's likely
to change in the future, plus it's inconsistent with other ways of
running/configuring happo).

I'm following instructions here to inspect and adjust jsdomErrors before
we log them:
https://github.com/jsdom/jsdom#virtual-consoles

First, I considered suppressing these logs completely. But that would
mean some issue go unnoticed. So truncation it is.