Allow plugins to inject css outside of rendering examples
I'm working on the happo-plugin-scrape plugin where I'm running into
performance issues related to css handling. The slow part is injecting a
large block of css collected in the scrape phase to the JSDOM instance.
Instead of doing that, we can allow the plugin to directly inject css,
avoiding the need to pass this css through the JSDOM instance.