Include tmpdir when inlining resources
In a create-react-app setup, webpack is configured with a loader that
allows you to import image files directly (e.g. `import img
from'./img.png'`) in your application code. The result of this is that
the file is copied (and fingerprinted) to the webpack output folder, and
imported as a (relative) path to the file. Since happo outputs things to
the os.tmpdir() folder, we need to look there when inlining image
resources.