Add project param
Now that happo.io accepts a `project` param for certain API endpoints,
we can start making use of it.

A happo.io account can have multiple projects. There's at least one,
default project. If you don't specify the project in config, we'll use
the default project as configured at happo.io. If the project is
configured, we'll associate reports/comparisons with that project. If
the project specified in .happo.js config does not exist, it will be
created the first time you use it.