Improve finding base commit of PR in Circle CI
Grabbing the latest commit from master is problematic. Lets say you
start working on a PR, based on commit sha "abc" from the master branch.
While you work on that PR, someone else merges a commit, "xyz" that adds
a bunch of happo examples to the master branch. If we run the happo
comparison on your PR, we'll see a bunch of deleted happo examples
because your commit didn't have the new examples.

To fix this, we can find the commit on master that the current branch is
based on. Found here:
https://stackoverflow.com/a/42562844