Make things work without a .happo.js config file
This will make it easier to use Happo out of the box. Instead of relying
on the user to fill in targets and apiKey/apiSecret, we can default the
values.

Having the api tokens defined as environment variables is a good thing
anyway, so this should simplify some documentation going forward and
make things a little more secure.