Wait for html from empty renders
When examples render nothing into the DOM, users end up seeing blank
screenshots. They can then choose to return a promise from the render
method in order to wait for content. As documented here:
https://github.com/happo/happo.io#asynchronous-examples

This is somewhat complex however. By providing a saner default, we can
likely help a few along the way. The new default involves waiting for
content to appear before we finish up rendering of that example.