Allow Happo to run in CI without pull request context
I want to make it easier to get happo up and running in CI. Right now,
the only way to make things work is to configure the build to run in
pull request environments (Circle CI and Travis have toggles for this).

After this change, jobs can be triggered

- on a branch (with or without PR context) - will trigger the double-run
- on the default branch (usually master) - will trigger a single report
for that commit sha

Going forward, I intend to make Happo.io post back a status on
individual reports as well as comparisons. I also might make use of
information from an optional pull request to contextualize a comparison.