Avoid unbound variable errors from ci scripts
Checking if a variable is empty is not enough to save us from it being
missing. This bash script will cause an error:

set -u
if [ -z "$FOO" ]; then
  echo 'Foo is missing'
fi

If we default the variable to an empty string, we can work around
unbound variable errors.

Also restructures the ci scripts so that all exported variables are at
the bottom in order to improve the overview.