Make Travis script work for push builds
The commit range we get from Travis is always present, even for push
builds. This means we can't trust it for grabbing the base commit sha.
Instead, do what the circleci script does, look up the commit on the
master branch which this commit is based on.

In the future, we'll have to change this to look at the default branch
instead of assuming master.