Improve stack traces with source maps
Instead of seeing line and column information from the built webpack
bundle, users are better helped with stack traces that refer to their
local files. Before this change, we were using `devtool:eval` which did
provide some help but it got line numbers wrong and wasn't great.

source-maps-support is a lib that helps turn regular stack traces into
sourcemapped ones using the source map referenced by the webpack bundle.
I had to use a few workarounds to get to a good state however:

- since stack traces are created by jsdom, they had `file://` prefixes
(that's how we load the js bundle into the dom). The `retrieveFile`
option passed to source-map-support helps us load the right file and
map.

- Due to how our WrappedError messed with `error.stack`, we weren't
getting source-map benefit directly. I had to change the implementation
slightly to make it less intrusive (leaving stack handling to
source-map-support), then recurse a little when logging.

- To make stacks easier to read, we strip out "file://<absolute path>"
to make paths relative to where you are.

An added benefit from changing the webpack config to development mode
is that we no longer use production/minified builds for dependencies
like react/react-dom etc.

I did a lot of trial-and-error to get to this point. Tried different
devtools for webpack, switched to the SourceMapDevToolPlugin with
different options passed in. I also tried having jsdom resolve the
source maps by using a localhost server to serve the bundle instead of
fetching from a file:// url. That was a fun little project but had lots
of downsides (plus I never got the source maps to resolve).