Enhance top-level stylesheets config
Instead of treating all stylesheets as global, add a `conditional` flag
to them. Any stylesheets with `conditional: true` will only be applied
if the local variant being rendered has `stylesheets: ['foo']`, where
`'foo'` is equal to an id set on the top-level stylesheet.

In short, this commit makes it possible to do this in .happo.js

stylesheets: [
  '/path/to/styles.css',
  {
    id: 'foo',
    source: '/path/to/foo.css',
    conditional: true,
  },
  {
    id: 'bar',
    source: '/path/to/bar.css',
  },
],

With this config, /path/to/style.css and /path/to/bar.css will be applied
to all examples. /path/to/foo.css will only be applied to variants that
have a `stylesheets: ['foo']` config, e.g.

{
  component: 'Button',
  variants: {
    small: {
      render: () => <Button small>Small button</Button>,
      stylesheets: ['foo'],
    }
  },
}

However, /path/to/bar.css and /path/to/style.css will also be applied to
Button-small. Any stylesheet (be it a path string or an object) without
`conditional: true` will always be applied. Stylesheets with `optional:
true` will be applied to variants with a `stylesheets` array referencing
the `id` of the stylesheet.