Isolate and manage tmpdir
Instead of throwing everything in os.tmpdir(), we now create a
sub directory for happo. It will be unique to the project dir (cwd), the
same way bundle files and entry files were previously produced. The
benefit from this is that we can more confidently know what temporary
files are used, plus we should be able to spin up a simple
static-serving web server in the tmp folder where you can inspect
examples.