Add `happo debug` command
This command will spin up a localhost server serving the contents of the
tmp folder where happo puts its files.

This rough first version will allow people to render all their examples
in a browser and get a summary of the results (number of errors, error
messages). The command won't generate the files for you, you will have
to run happo (either run or dev command) once first.

I moved rimrafing + mkdirp to the commands themselves so that files are
kept in the tmpdir after a run (loadUserConfig wasn't the right place
for this).