Attach assets as zip package instead of inlining
Inlining becomes tricky when the number of examples is big, since the
data is duplicated all over the place. If you for instance have 20
examples all using `/foo.png`, that image will be inlined in 20 places.

Happo workers are already able to handle zipped packages of assets (for
happo-plugin-storybook, that image will be inlined in 20 places.

Happo workers are already able to handle zipped packages of assets (for
happo-plugin-storybook), and we just need to make a minor change there.

The only tricky part after making this change will be to deal with fonts
on the workers. Inlining fonts makes them available right away, but now
we'll have to wait for them. I'll deal with that issue separately.