Allow payloads to be split into multiple chunks
This commit introduces a new option to RemoteBrowserTarget: `chunks`.
The default is 1.

If chunks is larger than 0, it will split up the payload (html snippets)
into multiple chunks that will be executed individually. This can help
users who have payloads that are too large for the server to handle. It
also has the added benefit of potentially running things in parallell,
leading to faster happo runs. This is all up to the happo.io server
however, parallellism isn't guaranteed.