Find assets relative to css files
If a css file has a relative url, we need to resolve it based on the
location of the css file. If a css file contains the following

// foo/bar/style.css
.foo {
  background-image: url('assets/a.png');
}

we need to resolve the asset at foo/bar/assets/a.png.

In order to do this, I needed to bring along the file path along the
way. It also involved allowing findCSSAssetPaths to return absolute file
paths in case the file path is there.