Include stylesheet-relative assets the right way
Since all css is served directly in the document on happo workers, asset
files need to be relative to the root instead of relative to the css
file location. We still need to resolve the file relative to the css
file, but the path inside the asset package need to be relative to the
root.

As an example, consider a css file located at
$project_root/path/to/styles.css:

.foo {
  background-image: url('assets/bg.jpg');
}

Before this change, the bg.jpg file would end up in the asset package as

  path/to/assets/bg.jpg

On happo workers however, all they see is the url in the css file,
`assets/bg.jpg`. We need to adjust our asset inclusion to always include
files relative to the root.