Fall back to resolving asset in public folders
There was one more case to handle for resolving assets. When a css file
has a relative reference to an asset, but that asset doesn't exist in
the same folder as the css file, we weren't including the asset in the
asset package. To fix this, we can add a fallback that will resolve
relative paths based on publicFolders only. Consider the following
example:

Css file /path/to/styles.css has this content:

.foo { background-image: url('foo.png') }

.happo.js has this content:

stylesheets: ['/path/to/styles.css'],
publicDirectories: ['/path/to', '/public'],

The file `foo.png` is located in /public/foo.png. When we tried to
resolve the file in the css file, we came up short, because it wasn't in
the same folder. But if we try looking in all public folders, we'll find
it.