Add warning log if config contains unknown key
To help users get the configuration right, I'm adding some logging when
we read the .happo.js file. If you try to use an unknown config key,
Happo will add a warning log.

I decided to go with the warning instead of failing because it's hard to
know for sure if the key is there for a real reason. And I want to
support new configs for old runs as well.