Allow a more granular retry logic for makeRequest
Not all requests benefit from being retried. By only retrying in certain
cases we can avoid unnecessary traffic to happo.io.

For instance, when fetching a report we don't care too much if it fails
or not. The only thing that happens if it fails is that the happo-ci
script will create a report for the base commit. Retrying here would
mean we would retry unnecessarily (since the response is often 404).