Handle potential off-by-one/rounding issue
From reports coming from a happo.io customer, I suspect the `chunks`
option for RemoteBrowserTarget isn't doing what it should in all cases.
While we haven't fully figured out what's wrong, a lot is pointing in
the direction of the chunks option:
- we're seeing 32 mysterious examples suddenly appearing
- there are 8 targets, each with 16 chunks

When I wrote this code I didn't think I'd be able to use floats for
Array.slice, but all the testing I did seemed to suggest it was working.