Don't silently fail when example has no render method
This commit fixes a confusing bug that had different symptoms in
different DOM providers. Basically, a happo file that looked like this:

export const foo = () => <div>Hello</div>;
export const bar = <div>Hello</div>;

- would not render any examples in puppeteer
- would silently ignore `bar` in jsdom

The silent failure in puppeteer was because we were trying to serialize
something that wasn't serializable. And then puppeteer silently returns
undefined. This bug:
https://github.com/GoogleChrome/puppeteer/issues/242

The ignore in jsdom was because we had code in place to ignore
render-less exports (some build tools add these).

The fix is to yell loudly as soon as we come across an example that's
misconfigured. I did change the example ignore logic a little, now
there's an assumption that any special exports added by build tools will
start with an underscore. This is potentially a breaking change, but I'm
willing to gamble a little and call this a patch. The extra exports I've
seen in the past all had underscores.