Make happo-ci script work when used with npx run
A bug was introduced in 3.13.0 where if you used npx to run happo-ci, it
would fail trying to run `happo start-job`. This is because we were
assuming a locally installed happo binary, but in the case of the npx
run, we have to use the folder where the current happo-ci job is
located.