Add support for remote rendering
By default, examples are pre-rendered in jsdom (or puppeteer when the
happo-plugin-puppeteer plugin is used). There are a few reasons we do
this:
- easier to surface errors to the user
- slightly better performance (although it was a long time since I
measured)
- no javascript code is shipped over the wire (instead, only html+css)

In the majority of cases, this approach is just fine. But recently I've
started coming across more and more cases where pre-rendering is making
things difficult. For custom web elements using the shadow dom,
pre-rendering is pretty much not working. To mitigate that, I'm
introducing a `prerender` configuration option. The default will be
`true`. By setting it to false, you leave rendering up to the happo.io
workers.