Add ability to post Happo statuses as comments
When using Github Enterprise, posting statuses from happo.io to the
local GHE installation is tricky. Github support recommends either
opening up their firewall to accept outside traffic on certain API
endpoints, or installing a local proxy that happo.io can talk to
somehow.

Both of those solutions require quite a bit of work in terms of
convincing ops people that this is a safe thing, plus configuring things
(on both ends). As a less heavy solution to getting the status back to a
PR from happo.io, I'm now exploring posting a comment with a "live"
happo status image.

The image is "live" in the sense that if a report first has diffs, then
someone accepts it, the image will (on page refresh) show the updated
state.

The downside of posting comments is that things get a little chatty.
Each time you push a commit and trigger CI, happo will post a comment.
And for now at least, the old comments will still be there.