Add workaround for old happo-plugin-puppeteer
This commit will ensure we don't have to make the next version bump a
major one. Since happo-plugin-puppeteer is missing a call to
happoProcessor.init({ targetName }), we were not properly filtering
examples. We can work around this for the time being by making the same
call in `next()` (like we've done before). At the next major version
bump we can remove this conditional block.

The only difference between the workaround and the main code path is
that the workaround does not support setting `targets` on an example.