Allow setting `maxHeight` for browser target
The default cutoff point on happo workers is 5000px. Any example larger
than that will be cropped. Now that we have full-page support, this
number is starting to feel a little small. But instead of globally
bumping it I'm making a change to RemoteBrowserTarget that will allow
you to set a custom maxHeight.