Improve finding the root node
Fixes #95.

Before, when an example renders to multiple DOM nodes, we would just
include one of them in the html snapshot. In many cases, this led to
empty/blank screenshots. Consider this html:

```html
<body>
  <div id="happo-root"><div></div></div>
  <div>
    I'm in a portal!
  </div>
</body>
```

With the old findRoot() implementation, only the inner div of
`#happo-root` would be included in the html snapshot. With the new
implementation, both elements would be included:

```html
<div id="happo-root"><div></div></div>
<div>
  I'm in a portal!
</div>
</html>

This change is a little delicate as we have to be careful about
including too much in the html snapshot. For instance, if a React
component only renders to a portal, we don't want to include the empty
root:

this
```html
<body>
  <div id="happo-root"></div>
  <div>
    I'm in a portal!
  </div>
</body>
```

should result in
```html
<div>
  I'm in a portal!
</div>
```

Why is this important? Because we want to keep screenshots as small as
possible, and by including outer divs we (usually) end up with page-wide
screenshots. If we have the portal render in a narrow container, the
empty `#happo-root` div would cause the screenshot to be full-width.

this
```
<body>
  <div id="happo-root"></div>
  <div style="width: 100px">
    I'm in a portal!
  </div>
</body>
```

should produce a screenshot that is 100px wide. If we were to include
`#happo-root` here, the screenshot would be as wide as the viewport is
(unless there's some css preventing default div styling to apply to
`#happo-root`).