Throw if asset path cannot be resolved
I'm setting up happo for a project and am running into an error:

> entry name must be a non-empty string value

This is coming from the archiver package, and I think it can happen if
the `name` we pass to it is falsey. But it is hard for me to understand
where the problem might be in my setup, since this is all of the
information I have to work from.

In order to produce a more actionable error, I think we should throw
here.