Ignore empty <img> src attributes
When finding assets in the html snapshots, we were not properly handling
empty src attributes. This led to a cryptic error down the line:

  Error: entry name must be a non-empty string value

We can simply ignore empty src to fix this. While I was trying to
reproduce the issue, I wrote a test case for css asset finding as well,
and I decided to keep it around even if it didn't cause any trouble.