Replace srcset parsing regex with parse-srcset package
We found a bug where srcsets with multiple URLs and mixed descriptors
would be parsed incorrectly, causing the first URL without a descriptor
to include a trailing comma. Instead of relying on this regex, I found
what appears to be a good package for parsing this string according to
the spec.

This is a port of this PR: https://github.com/Galooshi/happo/pull/229