Find assets from inline styles
Fixes #100

I have a snapshot that renders the following element:

```html
<div class="background_3l6wgj-o_O-large_usp5di" style="background-image: url(19674e0d6b42a252d31d6290dc406e1b.png);" role="img" aria-label="image"></div>
```

When happo takes a screenshot of this element, the image is missing.
When I view the snapshot source, I see the network request for this
image 404.

I believe that when the assets package is prepared, it is looking at
global CSS files, snap payloads, and things added to the public folders.

https://github.com/happo/happo.io/blob/b4ceb19bec/src/domRunner.js#L90-L94

The snapPayloads.assetPaths are added to the assets package here:

https://github.com/happo/happo.io/blob/b4ceb19bec/src/prepareAssetsPackage.js#L82-L86

So I believe the relevant part here is how snapPayloads.assetPaths is
generated. That happens in this loop:

https://github.com/happo/happo.io/blob/b4ceb19bec/src/processSnapsInBundle.js#L22-L28

Which calls this:

https://github.com/happo/happo.io/blob/b4ceb19bec/src/JSDOMDomProvider.js#L68-L70

Which then calls this:

https://github.com/happo/happo.io/blob/b4ceb19bec/src/browser/processor.js#L142-L174

I think the key line here is the call to findAssetPaths():

https://github.com/happo/happo.io/blob/b4ceb19bec/src/browser/processor.js#L167

Which is this function:

https://github.com/happo/happo.io/blob/b4ceb19bec/src/findAssetPaths.js#L9-L21

It looks like findAssetPaths will look for img src or img srcset, but
not other types of URLs, like in inline styles.

We can fix this by looking for elements with style attributes and
running those through findCSSAssetPaths.

It looks like findCSSAssetPaths takes a source option, which I wasn't
really sure what to do about or if it is needed here, so I left it out.
This function also gives an assetPath and a resolvePath, which I wasn't
really sure what the difference was and which one we want here. Looking
at the other test cases, it seems like resolvePath is more consistent
with what other things are doing.