Move cleanup git-checkout last
This will make it so that the happo compare call is made before cleaning
up and checking out the right commit. The compare call is what causes
the github status to get updated, so calling compare earlier will give
users feedback on the happo status sooner.

Also updated the test for happo-ci to also verify git calls, plus made
sure that we use the HAPPO_GIT_COMMAND env variable so that tests can
mock the call.