Only run CI reset after previous report is generated
This commit fixes two things. First of all, it improves performance a
little by only resetting git and node_modules when we know we've checked
out the previous commit. Secondly, this change fixes a bug where the
`happo compare` call would be run on the previous commit instead of the
current one in some cases.

Running `happo compare` on the previous commit would mean certain errors
would be hard to debug (we might be using an old happo.io module in the
previous commit).