Add ability to deep-compare reports
This commit will introduce a new configuration option: compareThreshold.
The purpose with this option is to allow people to deep-compare reports.
A deep comparison involves looking at actual pixels to decide whether or
not the two snapshots are okay or not.

I'm basically bringing over the happo-deep-compare module into this
project. It was an experiment I did for a while separately from the
happo.io module, and now it's time for its graduation.
https://github.com/happo/happo-deep-compare.

There is one key differences between the happo-deep-compare module and
the new implementation:We're not doing LCS (longest common subsequence)
diffing. This will make the deep-compare faster which will benefit
users. A side-effect of this is that we can shortcut execution if the
images are known to be of different size (not yet implemented but will
come in a subsequent commit).