Apply compareThreshold to individual pixels
Applying the compareThreshold on the average for the full diff risked
letting large images with small but very distinct diffing portions
through. Instead of using the average, we can do what @lencioni
suggested in
https://github.com/happo/happo.io/pull/109#issuecomment-530897774, apply
the threshold to individual pixels.

There are a few benefits from this approach. First, it makes it slightly
easier to reason about the threshold. Secondly, it makes the diffing
process faster since we can stop when we encounter a diffing pixel above
the threshold.